:truck: Move Trick to cards folder
