+++ server/src/test/java/com/coincoinche/engine/MoveBiddingTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 import com.coincoinche.engine.cards.Suit;
+import com.coincoinche.engine.cards.Trick;
 import com.coincoinche.engine.contracts.Contract;
 import com.coincoinche.engine.contracts.ContractFactory;
 import com.coincoinche.engine.game.GameResult;
